Rename resource to representor
