+++ lib/representor.rb
@@ -1,4 +1,4 @@
-require 'representor/resource_builder'
+require 'representor/representor_builder'
 
 ##
 # A canonical interface for introspecting a hypermedia message. A builder is used to construct the object
@@ -6,7 +6,7 @@ require 'representor/resource_builder'
 # generic hypermedia media types.
 #
 # @example
-#   my_resource = Representor.new do |builder|
+#   representor = Representor.new do |builder|
 #     builder.add_meta_link(:profile, 'http://alps.io/schema.org/ItemList')
 #
 #     builder.add_transition(:self, 'http://example.com/something')
@@ -22,37 +22,37 @@ require 'representor/resource_builder'
 #
 #     builder.add_attribute(:about, 'A list of people.')
 #
-#     builder.add_resources(:people) do |resource_builder|
-#       resource_builder.add_transition(:self, 'http://example.com/person/1')
-#       resource_builder.add_attribute(:first_name, '')
-#       resource_builder.add_attribute(:last_name, '')
+#     builder.add_representors(:people) do |representor_builder|
+#       representor_builder.add_transition(:self, 'http://example.com/person/1')
+#       representor_builder.add_attribute(:first_name, '')
+#       representor_builder.add_attribute(:last_name, '')
 #     end
 #
-#     builder.add_resources(:people) do |resource_builder|
-#       resource_builder.add_attribute(:first_name, '')
-#       resource_builder.add_attribute(:last_name, '')
-#       resource_builder.add_transition(:self, 'http://example.com/person/2')
+#     builder.add_representors(:people) do |representor_builder|
+#       representor_builder.add_attribute(:first_name, '')
+#       representorbuilder.add_attribute(:last_name, '')
+#       representor_builder.add_transition(:self, 'http://example.com/person/2')
 #     end
 #   end
 #
-#   self_transition = resource.transitions[:self]
+#   self_transition = representor.transitions[:self]
 #   self_transition.uri # => 'http://example.com/something'
 #
-#   search_transition = resource.transitions[:search]
+#   search_transition = representor.transitions[:search]
 #   search_transition.parameters.keys # => [:name]
 #
-#   create_transition = resource.transitions[:create]
+#   create_transition = representor.transitions[:create]
 #   create_transition.attributes.keys # => [:first_name, :last_name]
 #
-#   resource.attributes[:total_count] # => 2
+#   representor.attributes[:total_count] # => 2
 #
-#   people = resource.resources[:people]
+#   people = representor.representors[:people]
 #   person = people.first
 #   person.first_name # =>
 #
 class Representor
-  def initialize(resource_struct = nil)
-    builder = ResourceBuilder.new(resource_struct)
+  def initialize(representor_struct = nil)
+    builder = RepresentorBuilder.new(representor_struct)
     yield builder if block_given?
     @internal_struct = builder.to_struct
   end
@@ -65,21 +65,21 @@ class Representor
   end
 
   ##
-  # Any meta data about the resource.
+  # Any meta data about the representor.
   #
   def meta_data
     @meta_data ||= @internal_struct[:meta_data]
   end
 
   ##
-  # The attributes of the resource.
+  # The attributes of the representor.
   #
   def attributes
     @attributes ||= @internal_struct[:attributes] || {}
   end
 
   ##
-  # The transitions available for the resource.
+  # The transitions available for the representor.
   #
   # @return [Hash] The transition keyed by relation type.
   def transitions
@@ -87,10 +87,10 @@ class Representor
   end
 
   ##
-  # The separate resources embedded in the current resource.
+  # The separate representors embedded in the current representor.
   #
-  # @return [Hash of Representor] Any embedded resources keyed by name.
-  def resources
-    @resources ||= @internal_struct[:resources] || {}
+  # @return [Hash of Representor] Any embedded representors keyed by name.
+  def representors
+    @representors ||= @internal_struct[:representors] || {}
   end
 end
